+++ dev-packages/e2e-tests/test-applications/create-remix-app-v2/instrument.server.cjs
@@ -15,7 +15,4 @@ Sentry.init({
   environment: 'qa', // dynamic sampling bias to keep transactions
   dsn: process.env.E2E_TEST_DSN,
   tunnel: 'http://localhost:3031/', // proxy server
-  // In the orchestrion variant, the channel-based Remix integration replaces the default
-  // OpenTelemetry one (same `Remix` name, so it overwrites the default).
-  integrations: injectOrchestrion ? [Sentry.remixChannelIntegration()] : [],
 });
