+++ packages/remix/src/server/sdk.ts
@@ -6,7 +6,7 @@ import { DEBUG_BUILD } from '../utils/debug-build';
 import type { RemixOptions } from '../utils/remixOptions';
 import { instrumentServer } from './instrumentServer';
 import { httpIntegration } from './integrations/http';
-import { remixIntegration } from './integrations/opentelemetry';
+import { remixIntegration } from './integrations';
 
 /**
  * Returns the default Remix integrations.
